[add]  global shortcut-key of launching poker
